+++ src/configuration.rs
@@ -1,32 +1,35 @@
 use std::sync::RwLock;
+use Either::Left;
+use tokio_util::either::Either;
+use tokio_util::either::Either::Right;
 
 #[derive(Debug, Clone)]
 #[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
 pub struct Configuration {
     #[cfg_attr(feature = "serde", serde(serialize_with = "ser_addr", deserialize_with = "de_addr"))]
-    pub addr: Result<String, &'static str>,
+    pub addr: Either<String, &'static str>,
     pub connect_sec: u64,
     pub idle_sec: u64,
 }
 
 #[cfg(feature = "serde")]
-fn ser_addr<S>(addr: &Result<String, &'static str>, serializer: S) -> Result<S::Ok, S::Error> where S: serde::Serializer {
+fn ser_addr<S>(addr: &Either<String, &'static str>, serializer: S) -> Result<S::Ok, S::Error> where S: serde::Serializer {
     match addr {
-        Ok(a) => serializer.serialize_str(a),
-        Err(a) => serializer.serialize_str(a),
+        Left(a) => serializer.serialize_str(a),
+        Right(a) => serializer.serialize_str(a),
     }
 }
 
 #[cfg(feature = "serde")]
-fn de_addr<'de, D>(deserializer: D) -> Result<Result<String, &'static str>, D::Error> where D: serde::Deserializer<'de> {
+fn de_addr<'de, D>(deserializer: D) -> Result<Either<String, &'static str>, D::Error> where D: serde::Deserializer<'de> {
     use serde::Deserialize;
-    Ok(Ok(String::deserialize(deserializer)?))
+    Ok(Left(String::deserialize(deserializer)?))
 }
 
 impl Default for Configuration {
     fn default() -> Self {
         Self {
-            addr: Err("localhost:0"),
+            addr: Right("localhost:0"),
             connect_sec: 30,
             idle_sec: 60,
         }
@@ -34,7 +37,7 @@ impl Default for Configuration {
 }
 
 static CONFIG: RwLock<Configuration> = RwLock::new(Configuration {
-    addr: Err("localhost:0"),
+    addr: Right("localhost:0"),
     connect_sec: 30,
     idle_sec: 60,
 });
@@ -55,8 +58,8 @@ pub fn get_config() -> Configuration {
 pub fn get_addr() -> String {
     let c = CONFIG.read().unwrap();
     match &(*c).addr {
-        Ok(a) => a.clone(),
-        Err(a) => a.to_string(),
+        Left(a) => a.clone(),
+        Right(a) => a.to_string(),
     }
 }